+++ src/form-elements/FormElementBarcodeScanner.tsx
@@ -164,6 +164,7 @@ function FormElementBarcodeScanner({
                 value={value}
                 validationMessage={validationMessage}
                 lookupButtonConfig={element.lookupButton}
+                overrideRequiredMessage={undefined}
               />
             </div>