+++ src/form-elements/FormElementDateTime.tsx
@@ -124,6 +124,7 @@ function FormElementDateTime({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            overrideRequiredMessage={undefined}
           />
         </div>