+++ src/form-elements/FormElementEmail.tsx
@@ -83,6 +83,7 @@ function FormElementEmail({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            overrideRequiredMessage={undefined}
           />
         </div>