+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BPay handling to NSW GovPay provider
 - `formsAppService.getFormSubmissionMetaList()`
 
+### Changed
+
+- required logic for `lookupButton` form element
+
 ## [10.1.2] - 2026-05-07
 
 ### Added